+++ CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.0.9] / 2022-02-14
+## [0.0.9] / 2022-03-23
+### Added
+- Add CreateOrSelectPanel
+### 2022-02-14
 - Update AutodeskExtension - Add IsApplicationActive 
-- 2022-02-14
 - Update Build
 - Clear App